Move HTTP basic auth logic from WebsiteMode to Web, so it applies to all modes
